requirements.txt
flask==3.1.0
deepface==0.0.93
numpy>=1.14.0
pillow>=5.2.0
flask-cors>=4.0.1
requests>=2.27.1